requirements.txt
streamlit
requests
python-dotenv
Pillow
supabase
urllib3
backoff
